test: Use vows to run existing tests
Renamed the qualifying `.js` files to `.vows.js` to simplify filtering.
