--- test/dom/level3.js
@@ -1,8 +1,0 @@
-var wows = require('vows');
-var DOMParser = require('../../lib/dom-parser').DOMParser;
-
-// Create a Test Suite
-wows.describe('XML Namespace Parse').addBatch({
-	"test":function(){}
-    //see namespace.js
-}).export(module); // Run it
