+++ test/parse/unclosedcomment.vows.js
